+++ lib/plausible_web/controllers/api/stats_controller.ex
@@ -36,9 +36,10 @@ defmodule PlausibleWeb.Api.StatsController do
   end
 
   defp fetch_top_stats(site, %Query{filters: %{"goal" => goal}} = query) when is_binary(goal) do
+    total_filter = Map.put(query.filters, "goal", nil)
     prev_query = Query.shift_back(query)
-    unique_visitors = Stats.unique_visitors(site, %{query | filters: %{}})
-    prev_unique_visitors = Stats.unique_visitors(site, %{prev_query | filters: %{}})
+    unique_visitors = Stats.unique_visitors(site, %{query | filters: total_filter})
+    prev_unique_visitors = Stats.unique_visitors(site, %{prev_query | filters: total_filter})
     converted_visitors = Stats.unique_visitors(site, query)
     prev_converted_visitors = Stats.unique_visitors(site, prev_query)
     completions = Stats.total_events(site, query)